Add a test helpers to repeatedly try a sub-test until it succeeds
These helpers are similuar to assert.Eventually() and require.Eventually() but
take a sub-test instead of just a boolean predicate. This allows to perform
multiple checks in the same eventually call (useful to wait for an event and
then check multiple values). It also allows gives more useful output on
failures as the actual failure reason is shown.