MySQL schema import: use split function from go-libs module
MySQL supports sending multiple statements separated by `;` to the server.
However, when writing schema files containing stored functions, in order to be
able to import them using the regular mysql command line client, you have to
make use of the `DELIMITER` keyword implemented by that client and not
supported by the server. To be able to import such a file, splitting it into
statements while repecting the delimiter keyword has to be reimplemented. This
was done in the go-libs module and this commit makes use of that functionality.