Initial support for Icinga Notifications
Inspired by the existing code for the Icinga DB, support for Icinga
Notifications was added. Thus, there might be some level of code
duplication between those two.

The custom Icinga 2 configuration was sourced from the Icinga
Notifications repository, but edited to not being parsed as a faulty Go
template.

To receive notifications, IcingaNotificationsWebhookReceiver can be used
to launch a host-bound web server to act on Icinga Notification's
Webhook channel.