Replace Icinga/go-libs with Icinga/icinga-go-library
To sunset the go-libs project[0], the last externally used function,
MysqlSplitStatements, was moved to the icinga-go-library[1]. This change
is now reflected here as well, including bumping the go version in
go.mod to match the version required of the icinga-go-library.

This bigger Go version jump resulted in slightly reformatting the go.mod
file, splitting dependencies and transitive dependencies.

[0]: https://github.com/Icinga/go-libs/issues/4
[1]: https://github.com/Icinga/icinga-go-library/pull/115